refactor: refine approach to patching `tarfile` to support filters
