+++ agent/tox.ini
@@ -18,5 +18,5 @@ deps =
 commands =
     {envbindir}/python setup.py develop
     {envbindir}/python -m black --check setup.py testflinger_agent
-    {envbindir}/python -m flake8 setup.py testflinger_agent --exclude testflinger_agent/tarfile_cff0a2d.py
+    {envbindir}/python -m flake8 setup.py testflinger_agent --exclude testflinger_agent/tarfile_patch.py
     {envbindir}/python -m pytest --doctest-modules --cov=testflinger_agent testflinger_agent
